[Xamarin.Android.Build.Tasks] Dont clean the designtime folder
Context https://bugzilla.xamarin.com/show_bug.cgi?id=60880

Turns out the design time build in Visual Studio does not
clean up after itself.. ever. So we really should not be
deleting the directory on a `Clean`.

This commit move that code into is own internal target
and updates the tests to reflect the changes.