[tests] fixed FrameworkLibDirectory on Windows (#1077)
When I originally setup `FrameworkLibDirectory` on Windows,
`setup-windows.exe` was the method for running these tests. Now that
`xabuild.exe` is used instead, we need this directory to also possibly
be used from within the build tree as when `IsUnix` is true.

Changes:
- `FrameworkLibDirectory` will return locations within the build tree on
windows
- A few places that used to have an `IsUnix` check such as
`AndroidMSBuildDirectory`, it made more sense to do a `Directory.Exists`,
and then fallback to using `FrameworkLibDirectory`
- Fixed the `CrossCompilerAvailable` method on Windows
- The `BuildAotApplication` test has a `Regex` that won't work on
Windows due to shortened paths. It looked best to skip this assertion.

This short fix a variety of tests on Windows, and hopefully un-ignore
many tests as well.