[xabuild] support <PackageReference />
The `NetStandardReferenceTest` has been failing on Windows due to its
use of `<PackageReference />`.

Windows appears to need to import two files:
- C:\Program Files (x86)\Microsoft Visual Studio\2017\Enterprise\MSBuild\Microsoft\NuGet\15.0\Microsoft.NuGet.props
- C:\Program Files (x86)\Microsoft Visual Studio\2017\Enterprise\MSBuild\Microsoft\NuGet\15.0\Microsoft.NuGet.targets

Luckily we can set two properties to specify these paths:
`$(NuGetProps)` and `$(NuGetTargets)`. I think this is only a problem on
Windows, as these files aren't present on macOS.