[Xamarin.Android.Build.Tasks] ProGuard enclosing char on windows (#1068)
I noticed on Windows, the `Desugar(True, True, True)` test was
failing with:

```
java.io.IOException: Can't read
[E:\A\_work\1\s\bin\TestDebug\temp\Desugar(True,True,True)\obj\Release\android\bin\classes\classes.zip(!META-INF\MANIFEST.MF)]
(No such file or directory)
```

However, running locally proved the file does exist.

Somehow #1049 introduced a case where proguard is failing due to the
combination the new `(!META-INF/MANIFEST.MF)` filter and the
parameters passed to proguard in this test.

Looking into it, we had an odd form of how the paths were being
passed on Windows using a combination of single and double-quotes:

```
-injars "'path1(!META-INF/MANIFEST.MF)';'path2(!META-INF/MANIFEST.MF)'"
```

The way these paths were passed was not working on Windows after
adding this new filter--but only for this `Desugar` test... Most
cases were working, such as our tests around multi-dex and proguard.

It worked correctly on Windows when specified as:

```
-injars "path1(!META-INF/MANIFEST.MF)";"path2(!META-INF/MANIFEST.MF)"
```

So keep things consistent, I refactored the code to use `"` on
Windows and `'` elsewhere. This also resolved the issues in the
`Desugar` test on Windows, for which I also added an additional
assertion.