[enumification] fix RemoteConference.ConnectionProperties (issue #1087)
See the github comments on the issue details.

To solve the issue, we introduce a "fake" mapping fixup in enumification-helpers.

Use "force-api-level" mapping information to override the correct API Level
information to generate modified map(.ext).csv.

methodmap.csv is also fixed to (1) remove old mapping for 24 and (2) change
the mapping for 25 down to 24 too.