[tests] fixed several tests on Windows (#1086)
Changes:
- `CheckResourceDesigner*` for F# can't pass on Windows due to the
MSBuild project API sorting. F# relies on files in the `fsproj` in a
certain order, which cannot be achieved if the MSBuild API always does
its own sorting. I feel it is best to skip these tests on Windows for
now.
- `BuildExternalJavaLibrary` should just use `$(MSBuildExtensionsPath)`
since `$(MonoDroidInstallDirectory)` is not set on Windows. It also
needed to apply a timestamp to the new reference, or the project file
wasn't saving on Windows. The last fix is using `Path.Combine`.
- `BuildAppCheckDebugSymbols` needs to set `DebugType` to `portable` on
Windows in order to produce pdb files
- `BuildAppWithManagedResourceParserAndLibraries` was consistently
taking ~9 seconds on the VSTS build agent. I was also hitting the 260
max path limit on my machine, so I shortened the directory name.
- `CheckBuildIdIsUnique` needed to take into account that `mdb` files
are generated on Windows
- `CheckLibraryImportsUpgrade` needed a change to force the project file
to re-save on Windows. I think this is happening because the timestamps
have down to the millisecond on Windows, so Windows more thoroughly
enforces the save-skipping logic. I also needed to change one assertion
on Windows, since `IncrementalClean` is deleting a directory on `Build`.
This is fine.
- `KnownPackages` had the wrong `<Reference />` name for
`Xamarin.Android.FSharp.ResourceProvider.Runtime`
- Added a `ToString` implementation for `BuildItem` since it greatly
improved the debugging experience. You can mouseover an item and see the
value of `Include()`.