[tests] Fix apk sizes input filename
Previous commit
https://github.com/xamarin/xamarin-android/commit/dd42c9a41fac43a96b001e374c7a691aff3284c8
tried to fix apk sizes input filename, as it was broken for some
time. I overlooked that `Package` is part of the item group though and
not a property group as I thought. Thus we got all measurement mixed,
so last one won and thus we have wrong numbers for Mono.Android
test. This patch fixes that.