[tests] Fix apk sizes input filename (#1117)
Previous commit dd42c9a4 tried to fix apk sizes input filename,
as it was broken for some time. I overlooked that
`%(TestApk.Package)` is part of the item group though and not a
property group as I thought. Thus we got all measurement mixed,
so the last one won and thus we have wrong numbers for the
Mono.Android test. This patch fixes that.