Fix bundling of assemblies in NDK with unified headers (#1105)
Context: https://developercommunity.visualstudio.com/content/problem/166029/building-xamarinandroid-application-with-bundleass.html

NDK 14 introduced experimental support for so-called "unified headers" which
replace the per-platform C header files, NDK 15 deprecated the old headers and
NDK 16 removed them completely. Xamarin.Android has been building itself with
NDK configured to use unified headers for a while now, but we missed one piece
of code which is affected by the change - mkbundle.

This commit adds code to MakeBundleNativeCodeExternal task to make it possible
to build code generated by mkbundle on systems with NDK that lacks the
per-platform header files. This is done in a backward-compatible way but if your
NDK has both the old and the new headers, the latter will be used.

Additionally, minimum API versions to 14 for 32-bit targets (forward
compatibility with NDK r16)