[Xamarin.Android.Build.Tasks] ProGuard/MultiDex fix parentheses in path
Fixes: #1135
Context: #1068

When making changes in #1068, I added a unit test that builds an
application with both spaces and parentheses in its path. However, this
was not enough to test everything; the test also needed an
`AndroidJavaLibrary` build item with spaces and parentheses in its name!

Changes:
- Made a `Hello (World).jar` file, that has a single class in it
- Added the jar file as a `AndroidJavaLibrary`, in
the `BuildApplicationWithSpacesInPath` test
- Reverted `Proguard.cs` and `CreateMultiDexMainDexClassList.cs` from
what I did in #1068 -- Dean tried to warn me
- Move the `(!META-INF/MANIFEST.MF)` filter expression outside the
single quotes
- For the `-injars` expression to work on *both* macOS and Windows,
I had to enclose the entire expression in double-quotes

PR #1068 was a good attempt at trying to cleanup the code, but I could
not get the manifest filter to work along with a path including a
parentheses.