[Xamarin.Android.Build.Tasks] ProGuard/MultiDex fix parentheses in path (#1137)
Fixes: https://github.com/xamarin/xamarin-android/issues/1135
Context: commit 2c2d47c4

When making changes in 2c2d47c4, I added a unit test that builds an
application with both spaces and parentheses in its path. However,
this was not enough to test everything; the test also needed an
`@(AndroidJavaLibrary)` build item with spaces and parentheses in
its name!

Changes:

  - Made a `Hello (World).jar` file, that has a single class in it
  - Added the jar file as a `AndroidJavaLibrary`, in
    the `BuildApplicationWithSpacesInPath` test
  - Reverted `Proguard.cs` and `CreateMultiDexMainDexClassList.cs` from
    what I did in 2c2d47c4 -- Dean tried to warn me
  - Move the `(!META-INF/MANIFEST.MF)` filter expression outside the
    single quotes
  - For the `-injars` expression to work on *both* macOS and Windows,
    I had to enclose the entire expression in double-quotes

2c2d47c4 was a good attempt at trying to cleanup the code, but I
could not get the manifest filter to work along with a path including
a parentheses.