[Xamarin.Android.Build.Tasks] <Aapt /> task showing warnings as errors
Fixes: #1134
Context: 2135856e

Dean's changes in 2135856e were more accurate at picking out warnings
& errors, but if the `level` value from the `Regex` was blank, it was
counting the message as an error. The `level` value is matching against
the words `warning` or `error`, ignoring case.

I think the fix here is to count the message as a warning if the
`level` is blank.

I added a Regex test case of what was on #1134. I added another test case
that verifies a message with a blank `level` comes through the build output
as a `warning`, not giving an `APT0000` error.