[Xamarin.Android.Build.Tasks] Fix message for missing Platform error XA5207
Fixes #1136

Commit 7f45a6d reworked the GetJavaPlatformJar Task (and others)
to handle a NRE exception. However the error logging code in
MonoAndroidHelper.cs was not formatted correctly. As a result
the XA5207 error would be logged without any message.

This commit fixes up the LogError call to make sure the error
has a decent message. It also adds a unit test so that we can
catch this issue if it regresses in the future.