[Xamarin.Android.Build.Tasks] Rework Aapt log processing.
Context #1134

Aapt is pretty inconsistent with how it reports errors and
warnings. Warnings are written to stderr..

So we are reworking the output processing for aapt to hopefully
handle errors are warnings in a more consistent way. Rather than
process the output in realtime we will buffer it and process the
output once aapt has completed. This will allow us to check to
see if the required output file (R.java) was created. If it
wasn't then we can assume that the process failed.

We can then process the output knowing that we are expecting
and error. In the case where R.java does exist, we can assume
if we do find anything which looks like an error it is more
than likely a warning.