Sanitize Xamarin.Android.sln: fix several inconsistency bits.
Manually editing solution file is uncool and error prone, but since the
entire xamarin-android source tree is not designed cleanly it is impossible
to use IDEs to do the right job.

As stated at https://github.com/xamarin/xamarin-android/issues/1129 ,
we should eliminate this Xamarin.Android.sln as a source file and it should
be only generated, not directly committed to the tree. But to achieve that
we first need to find the right structure and right content.

So far, there were handful of issues that either missed or incorrectly
added solution items with inconsistent configurations.

These inconsistent parts were found by the solution content analyzer at
https://github.com/xamarin/xamarin-android/issues/1129#issuecomment-355484902