[Xamarin.Android.Build.Tasks] Always emit targetSdkVersion in the manifest.
Context #1163

We currently do not always emit `targetSdkVersion` in the
AndroidManifest.xml. If the user has `AndroidUseLatestPlatformSdk`
set (which is the default) then it definately doesn't get
written.

This can cause problems when a developer is trying to use
modern features. Because of the `targetSdkVersion` is not
specified it defaults to the `minSdkVersion` value.

So lets add a check to make sure we always emit this value.
We already have code which makes sure we do not overwrite
it if the user has specificed it. So all we need to do is
handle the case where they haven't added it.