Preparation for the new test runner
PR https://github.com/xamarin/xamarin-android/pull/1075 keeps rebuilding all the
Mono runtimes because the two files in this commit are part of the logic to
determine whether to use cached Mono or not. The result isn't cached, because
it's a PR builder, and so we keep waiting for 7h+ each time a smallest change is
made. This commit will cause Mono to be rebuilt and cached and, in turn, speed
up round-trip time for PR 1075