Better report on location of an Activity without default constructor (#1169)
Fixes: https://github.com/xamarin/xamarin-android/issues/1151

Each `Activity` must have a default constructor or it won't be possible to
instantiate it from the android manifest. Should the constructor be missing we
currently report the fact in a way that doesn't tell the developer where the
error occured. This is not very helpful.

This commit adds code to add the location of the offending source code, if
available, as well as changes the way we report the error - we no longer throw
an exception but instead define a new error code, XA4213, and report it using
the standard MSBuild logger.