[Xamarin.Android.Build.Tasks] Ignore the ObsoleteSdkInt Lint Check
Context #1171

There is a new test [ObsoleteSdkInt] which produces a warning
such as

	obj/Debug/android/src/mono/MonoPackageManager.java(85,6): warning XA0102:  Unnecessary; SDK_INT is always >= 10 [ObsoleteSdkInt]

So our MonoPackageManager.java file is the cause of this
particular warning [1]. I am not sure we can remove that
line of code yet, since our apps can run on older devices.
That said once the % share of devices for API's < 8 is 0,
we can probably remove that line of code completely and
re-enable this check.

[1] https://github.com/xamarin/xamarin-android/blob/master/src/Xamarin.Android.Build.Tasks/Resources/MonoPackageManager.java#L85