[Xamarin.Android.Build.Tasks] Fix Aapt task to emit warnings (#1178)
Fixes: https://github.com/xamarin/xamarin-android/issues/1134

Commit 23c5801 reworked `<Aapt/>` to emit warnings if the required
output file did exist. If it didn't we should emit errors.
There was one slight problem with this, we were checking the
wrong file. We should have been checking for the temp
`packaged_resources` file rather than `R.txt` (which will
always exist after an initial build).

This commit fixes that and adds a unit test. Note the in order
to create the senario we need to have a `$(TargetFrameworkVersion)`
which is `v8.0`. We currently only build `v1.0` and `v8.1` for
PR builds. So the test checks to see if that target framework is
available. If not the test will be ignored.