[Xamarin.Android.Build.Tasks] Ignore AndroidAsset items which are folders.
When adding a new subfolder in the IDE (VSForMac) we end up
with a folder which has a build action of AndroidAsset.
This causes the copy task to fail with

	Cannot copy <folder> to <destfolder>, as the source file doesn't exist",

This is because we assume that all items in AndroidAsset (and
AndroidResource) and files. It looks like this is not the case.
While a work around is easy .. simply remove the item. It does
require that the user manually update the csproj.

So this commit uses `Directory.Exists` to check to see if the
item is a directory or not. If it is we skip over it. Otherwise
its a file so we should copy.

We do this in `Files.cs` but also in `AndroidUpdateResDir.cs`.
The latter has been changed to output a list of the "Resolved"
files which will not include any folders.