Allow AndroidSdkInfo.SetPreferred*Path() to work (#17)
Commit ca5745db added `AndroidSdkInfo.SetPreferredAndroidNdkPath()`,
`AndroidSdkInfo.SetPreferredAndroidSdkPath()`, and
`AndroidSdkInfo.SetPreferredJavaSdkPath()`.

The problem is, these methods are not usable in all scenarios.

The purpose of these methods is to allow e.g. the IDE to allow the
developer to alter the preferred Android NDK, SDK, or JDK paths.
Implicit in this purpose is that the *existing* paths may be invalid.

Thus, if we assume that there is no Android SDK or JDK known by
`AndroidSdkInfo`, but "someone" knows of one and wants to use that in
the future, the obvious thing to do would be to call e.g.
`AndroidSdkInfo.SetPreferredAndroidSdkPath()`.

Therein lines the problem: these are *instance* methods, and the
`AndroidSdkInfo` constructor *requires* that *all* paths be valid,
otherwise the constructor may throw an `InvalidOperationException`.
This renders `SetPreferredAndroidSdkPath()` largely useless: you would
need to construct `AndroidSdkInfo` with a known instance in order to
set a preferred location, which is a tad backwards.

Turn `AndroidSdkInfo.SetPreferredAndroidNdkPath()`,
`AndroidSdkInfo.SetPreferredAndroidSdkPath()`, and
`AndroidSdkInfo.SetPreferredJavaSdkPath()` into `static` methods, so
that they may be called without constructing an `AndroidApiInfo`.