[Xamarin.Android.Build.Tasks] Ignore @(AndroidAsset) folders (#1190)
When adding a new subfolder in the IDE (VSForMac) we end up with a
folder which has a build action of `@(AndroidAsset)`. This causes the
copy task to fail with:

	Cannot copy <folder> to <destfolder>, as the source file doesn't exist",

This is because we assume that all items in `@(AndroidAsset)` (and
`@(AndroidResource)`) are files. It looks like this is not the case.
While a work around is easy -- simply remove the item -- it requires
that the user manually update the `.csproj`.

Instead, use `Directory.Exists()` to check to see if the item is a
directory and skip the item if it is. Otherwise its a file so we
should copy.

We do this in `Files.cs` but also in `AndroidUpdateResDir.cs`. The
latter has been changed to output a list of the "Resolved" files
which will not include any folders.