[Mono.Android] Expose android.graphics.Color as ColorObject and bind WallpaperColor.
It turned out that the criminal who generated _Color and ColorObject as
internal was *metadata fixup*, not generator itself. It was buried in
thousands of lines of `src/Mono.Android/metadata`.

Color is absolutely a technical debt that Xamarin.Android should clean up
and provide every related API in int properties and constants, but so far
add a workaround by renamed Color->ColorObject class.