[Xamarin.Android.Build.Tasks] Rework AndroidApkSigner to use .jar directly. (#1222)
Once again we hit an issue where the `apksigner.bat` file provided
with android build-tools was not functioning correctly. So like
we did with `dx.bat` and other tooling we should switch to using
the `apksigner.jar` file directly.

Additionally JDK 9 has removed support for the `java.ext.dirs`, which
all versions of the `apksigner.bat` use. As a result this tool will
not work with JDK 9. So its best to remove it now.

We have existing unit tests which check that the Apk is being
signed correctly. So not additional tests are required.