[Xamarin.Android.Build.Tasks] missing EmbeddedResource.LogicalName for __AndroidLibraryProjects__
Fixes #1235

There seems to be a bug in msbuild [1] where if you create
an item inside a task and the task is skipped, the metadata
is not created for the item.

So we get into a situation where our `__AndroidLibraryProjects__.zip`
does not contain the correct LogicalName. As a result our
extraction process cannot find the resource, and builds fail.

This commit spits out the code that created the zip file into
its own target `_CreateManagedLibraryResourceArchive`. This means
we can skip this new target it needed. However the target
which creates the item `_AddLibraryProjectsEmbeddedResourceToProject`
can now run for every build. This will ensure that the metadata is
always created.

Also added a unit test.

[1] https://social.msdn.microsoft.com/Forums/netframework/en-US/259271d8-a3fc-4a9b-9e3a-fecad8f6f63f/potential-bug-with-createitem-task-and-additionalmetadata?forum=msbuild