[Xamarin.Android.Build.Tasks] Generate error when JDK 9 is used
Fixes #1248

JDK 9 doesn't currently provide a robust and reliable user
experience. Until we can provide a reliable experience, we
should "fail early".

We already have a property for `$(LatestSupportedJavaVersion)`
which currently cannot be overridden. We also currently emit
a warning when the java version is greater than this property.
However with JDK 9 it seems a warning is not enough. So we
should change the warning to an error. We also need to make
`$(LatestSupportedJavaVersion)` writable so that it can be
overridden from the command line via.

	msbuild /p:LatestSupportedJavaVersion=9.0.0 /t:SignAndroidPackage TheApp.csproj