[Tasks] Add aot to the list of valid profilers
With this change it is possible to build an apk containing `aot`
profiler module (libmono-profiler-aot.so shared library) and profile
the app with it. It can enabled by setting `mono.debug.profile`
property to `aot` on the device or emulator.

Note that there is an issue with Java.Interop, where the app shutdown
is broken right now. We get unhandled exception and thus the shutdown
process doesn't finish, which means the profile output is not written
to the file. Hopefully I can fix that soon.