[Xamarin.Android.Build.Tasks] Stop IncrementalClean removing assemblies.
Fixes #1164
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=61108

The IncrementalClean target basically deletes a ton of stuff
if its not in the FileWrites group. Under certain conditons
this include assemblies in the `$(OutDir)`.
It mainly effects projects which use a shared `$(OutDir)`.
So it a third level dependency is copied to the `$(OutDir)`
but is not referenced by the main app project we end up with
IncrementalClean deleting the file. This then breaks packaging
since we expect the file to exist.
This causes problems with TFS because it seems to ALWAYS use
a shared `$(OutDir)`.
To work around this problem we just need to tell IncrementalClean
to ignore any dll in the `$(OutDir)`. Since we already have a
target in place to handle this we can just expand the list of
files it should be ignoring.